Rename integration tests to e2e tests
The Extension Host feature tests exercise full VS Code features end-to-end
(jack-in, connect, evaluate, session routing, etc.). Rename directories,
CI jobs, npm scripts, imports, env vars, and references from 'integration'
to 'e2e' to match their actual nature.
