+++ .circleci/config.yml
@@ -209,7 +209,7 @@ jobs:
         command: npm run unit-test
     - store_test_results:
         path: ~/calva/junit
-  test-integration:
+  test-e2e:
     docker:
     - image: cimg/clojure:1.11-browsers
     working_directory: ~/calva
@@ -227,12 +227,12 @@ jobs:
         command: sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys E88979FB9B30ACF2; sudo apt update && sudo apt install -y libnss3 imagemagick
     - run:
         name: Run Extension Tests
-        command: npm run integration-test
+        command: npm run e2e-test
     - store_test_results:
         path: ~/calva/junit
     - store_artifacts:
-        path: ~/calva/out/extension-test/integration/screenshots
-        destination: integration-screenshots
+        path: ~/calva/out/extension-test/e2e/screenshots
+        destination: e2e-screenshots
   test-smoke:
     docker:
     - image: cimg/clojure:1.11-browsers
@@ -505,7 +505,7 @@ workflows:
     - test-cljslib:
         requires:
         - build
-    - test-integration:
+    - test-e2e:
         requires:
         - build
     - test-smoke:
@@ -523,7 +523,7 @@ workflows:
         - eslint-check
         - test-grammar
         - test-cljslib
-        - test-integration
+        - test-e2e
         - test-smoke
         - test-smoke-sub-projects
         - test-ts-unit
@@ -579,7 +579,7 @@ workflows:
             ignore: /.*/
           tags:
             only: /^v\d+\.\d+\.\d+-?.*/
-    - test-integration:
+    - test-e2e:
         requires:
         - build
         filters:
@@ -617,7 +617,7 @@ workflows:
         - eslint-check
         - test-grammar
         - test-cljslib
-        - test-integration
+        - test-e2e
         - test-smoke
         - test-smoke-sub-projects
         - test-ts-unit
