+++ .circleci/workflows/build-test.yaml
@@ -12,7 +12,7 @@ jobs:
 
 - eslint-check:          {requires: [build]}
 - test-cljslib:          {requires: [build]}
-- test-integration:      {requires: [build]}
+- test-e2e:              {requires: [build]}
 - test-smoke:              {requires: [build]}
 - test-smoke-sub-projects: {requires: [build]}
 - test-ts-unit:          {requires: [build]}
@@ -23,7 +23,7 @@ jobs:
     - eslint-check
     - test-grammar
     - test-cljslib
-    - test-integration
+    - test-e2e
     - test-smoke
     - test-smoke-sub-projects
     - test-ts-unit
