+++ src/extension-test/e2e/suite/connection-isolation-test.ts
