+++ src/extension-test/e2e/suite/edit-replace-test.ts
