+++ src/extension-test/e2e/suite/file-output-destination-test.ts
