+++ src/extension-test/e2e/suite/highlight-test.ts
