+++ src/extension-test/e2e/suite/jack-in-and-connect-test.ts
