+++ src/extension-test/e2e/suite/jack-in-dependency-versions-test.ts
