+++ src/extension-test/e2e/suite/name-suffix-test.ts
