+++ src/extension-test/e2e/suite/project-utils-test.ts
