+++ src/extension-test/e2e/suite/util.ts
@@ -19,7 +19,7 @@ export const testDataDir = path.join(
   __dirname,
   ...['..', '..', '..', '..'],
   'test-data',
-  'integration-test'
+  'e2e-test'
 );
 
 export const isCircleCI = process.env.CIRCLECI === 'true';
@@ -49,11 +49,11 @@ export function sleep(ms: number): Promise<void> {
 }
 
 export function log(suite: string, ...things: any[]) {
-  console.log(`Integration testing, ${suite}:`, ...things);
+  console.log(`E2E testing, ${suite}:`, ...things);
 }
 
 export function showMessage(suite: string, message: string) {
-  void vscode.window.showInformationMessage(`Integration testing, ${suite}: ${message}`);
+  void vscode.window.showInformationMessage(`E2E testing, ${suite}: ${message}`);
 }
 
 export function getExecutablePath(executablePathMaybe: string) {
