+++ src/extension-test/e2e/suite/websocket-connect-test.ts
